requirements.txt
streamlit>=1.37
pandas>=2.2
numpy>=1.26
yfinance>=0.2.52
seaborn>=0.13
matplotlib>=3.9
plotly>=5.23
python-dotenv>=1.0
google-generativeai>=0.8